element_discovery/AST/requirements.txt
opencv-python
tensorboard
torchvision
natsort
numpy
scipy
tqdm >= 4.35.0
pytorch_model_summary
datetime
scipy
scikit-learn
seaborn
transformers
